requirements.txt
# ── Web Scraping ──────────────────────────────────────────────────────────────
playwright==1.44.0
playwright-stealth==1.0.6
httpx==0.27.0
fake-useragent==1.5.1

# ── HTML Parsing ──────────────────────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.2.2
cssselect==1.2.0

# ── NLP / Text Features ───────────────────────────────────────────────────────
langdetect==1.0.9
nltk==3.8.1

# ── ML ────────────────────────────────────────────────────────────────────────
xgboost==2.0.3
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
shap==0.45.0
imbalanced-learn==0.12.2

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm==4.66.4
rich==13.7.1
python-dotenv==1.0.1
joblib==1.4.2
pyyaml==6.0.1
click==8.1.7